# api/openapi.yaml
openapi: 3.0.0
info:
  title: BookFellas API
  description: A simple API of Book Fellas web application.
  termsOfService: http://swagger.io/terms/
  contact:
    name: Panos Lekos
    email: panaleko@ece.auth.gr
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
# servers:
# - url: https://virtserver.swaggerhub.com/BILONIO02_1/BookFellas/1.0.0
#   description: SwaggerHub API Auto Mocking
paths:
...
  /library/{id}:
    get:
      tags:
      - Library
      summary: Get a Library by id
      operationId: libraryIdGET
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Returns the Library
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Library'
        "404":
          description: Library not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: Library
    put:
      tags:
      - Library
      summary: Update a library by ID
      operationId: libraryIdPUT
      parameters:
      - name: id
        in: path
        description: The ID of the library to update
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Library'
        required: true
      responses:
        "200":
          description: The updated library
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Library'
        "404":
          description: Library not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: Library
  /library/{id}/book:
    get:
      tags:
      - Library
      summary: Get a list of books in a specific library
      operationId: libraryIdBookGET
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: A list of books in the library
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
                x-content-type: application/json
        "404":
          description: Library not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: Library
    put:
      tags:
      - Library
      summary: Updates a Library's book list. Use "add" or "remove" operation to update
        the list.
      operationId: libraryIdBookPUT
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
        required: true
      responses:
        "204":
          description: "Library updated, no response"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: Library
...
components:
  schemas:
...
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        publishedDate:
          type: string
          format: date
        isbn:
          type: string
      example:
        author: author
        isbn: isbn
        id: 0
        publishedDate: 2000-01-23
        title: title
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
    Library:
      required:
      - id
      - location
      - name
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier for the library
        name:
          type: string
          description: The name of the library
        location:
          type: string
          description: The location of the library
        established:
          type: string
          description: The date the library was established
          format: date
        booksCount:
          type: integer
          description: The total number of books in the library
        librarian:
          type: string
          description: The name of the head librarian
      example:
        established: 2000-01-23
        name: name
        librarian: librarian
        location: location
        id: 21
        booksCount: 2
...